# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Go Webapp Template API
  version: 0.4.0
  description: |
    Go Webapp Template identity and authentication API. This is the contract-first
    source of truth for the REST surface: Go transport types are generated
    from this file (see `make openapi-generate`), never the other way
    around.

    ## Versioning

    Two different version numbers appear here, and they are not the same
    concept:

    - The **URL major version** (`/api/v1`) is a compatibility boundary.
      It changes only for a breaking change — a removed or renamed field,
      a changed field type, incompatibly changed semantics, a restructured
      request or response. A new endpoint, a new optional response field,
      or a new optional query parameter are all backward compatible and
      stay on `/api/v1`.
    - **`info.version`** above is this document's own revision. It moves
      for any published change, including backward-compatible ones, while
      the URL major stays put.

    So `/api/v1` is expected to outlive many `info.version` bumps. When a
    genuinely breaking change does arrive, `/api/v2` is added alongside
    `/api/v1` rather than replacing it in place.

    Operational endpoints (`/health`) are infrastructure, not part of the
    public business API, and are deliberately unversioned: monitoring and
    orchestration probe a stable path that never moves with the API's
    compatibility boundary.
servers:
  - url: /
...
security: []
paths:
...
  /api/v1/auth/verify-email:
    post:
      tags: [auth]
      summary: Verify an email address using its one-time token
      operationId: verifyEmail
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '204':
          description: Email verified and token consumed
        '400':
          description: Malformed request or invalid, expired, consumed, or replaced token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: {code: invalid_email_verification, message: Email verification link is invalid or expired}
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate a refresh token for a new access/refresh token pair
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token rotated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Refresh token is invalid, reused, expired, or its account is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidRefreshToken:
                  value: {error: {code: invalid_refresh_token, message: Invalid refresh token}}
                accountDisabled:
                  value: {error: {code: account_disabled, message: Account is disabled}}
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    PayloadTooLarge:
      description: Request body exceeded the maximum allowed size
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: {code: request_body_too_large, message: Request body is too large}
    UnsupportedMediaType:
      description: Content-Type was missing or was not application/json
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: {code: unsupported_content_type, message: Content-Type must be application/json}
    TooManyRequests:
      description: |
        The caller exceeded a rate limit and should retry after the
        interval given in the Retry-After header.

        Every endpoint under /api is covered by a generic per-IP limit;
        the authentication endpoints carry stricter limits on top. The
        response is deliberately identical whichever limit was reached,
        so it reveals nothing about which one — in particular nothing
        about whether an account exists.
      headers:
        Retry-After:
          description: Whole seconds to wait before retrying, rounded up.
          required: true
          schema:
            type: integer
            minimum: 1
          example: 30
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: {code: rate_limit_exceeded, message: Too many requests}
    InternalServerError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: {code: internal_error, message: Internal server error}
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable error code.
              example: invalid_credentials
            message:
              type: string
              description: Human-readable presentation text. Never contains internal details.
              example: Invalid email or password
...
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 512
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    RefreshResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
          description: A newly rotated opaque credential; the previous one is now invalid.
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          format: int64